feat: Completion of the Simulation module with complete CRUD

+++ worldCup/src/main/java/com/worldcup/worldcup/modules/results/oficialMatches/repository/OficialMatchRepository.java
@@ -3,6 +3,6 @@ package com.worldcup.worldcup.modules.results.oficialMatches.repository;
 import com.worldcup.worldcup.modules.results.oficialMatches.entity.OficialMatch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OficialMatchRepository extends JpaRepository<OficialMatch, Long> {
+public interface OficialMatchRepository extends JpaRepository<OficialMatch, Integer> {
 
 }
